chore: consolidate crate::visitors::ast
This patch consolidates the `crate::visitors::ast` package into
a single, flat module.
